refactor: split AI sync script emitter
