+++ .sampo/changesets/1020-split-ai-sync-script-emitter.md
@@ -1,0 +1,7 @@
+---
+npm/@wp-typia/project-tools: patch
+npm/wp-typia: patch
+---
+
+Split the generated AI feature sync script source emitter into a focused runtime
+module while preserving compatibility exports and generated output.
